+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH3YN69V0SE309TVS2ZT2M1E"
+        "01KH3ZN3B1BS85SJ2EA65QF9JW"
     )
     r"""The unique identifier of the data source"""