+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH3YN69Q3QDA6SN74GQFTJX6"
+        "01KH3ZN3AZ90YETGV9TM6MKXFG"
     )
     r"""The unique identifier of the data source"""