+++ packages/orq-rc/RELEASES.md
@@ -5479,3 +5479,13 @@ Based on:
 - [python v4.3.0-rc.48] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.48] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.48 - packages/orq-rc
+
+## 2026-02-10 14:37:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.709.3 (2.814.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.49] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.49] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.49 - packages/orq-rc